% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_simple_rim.R
\name{add_simple_rim}
\alias{add_simple_rim}
\title{Remove Numbers from a Column in a Data Frame}
\usage{
add_simple_rim(data, column_name)
}
\arguments{
\item{data}{A data frame containing the column to be modified.}

\item{column_name}{The name of the column from which numbers should be removed.}
}
\value{
A data frame with numbers removed from the specified column.
}
\description{
This function takes a data frame and the name of a column as input. It removes numbers from the specified column using regular expressions and returns the modified data frame.
}
\examples{
\dontrun{
data <- data.frame(Name = c("T1c", "S9", "S2"))
modified_data <- remove_numbers(data, Name)
}

}
