% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_boxplots.R
\name{all_boxplots}
\alias{all_boxplots}
\title{Make boxplots from all numeric-categrocial pairs}
\usage{
all_boxplots(data)
}
\arguments{
\item{data}{A dataframe or tibble.}
}
\value{
A list of ggplot objects.
}
\description{
Make boxplots from all numeric-categrocial pairs
}
\examples{
p <- all_boxplots(guinea[[3]])
print(p)
}
