% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_missing.R
\name{count_missing}
\alias{count_missing}
\title{A function to count missing values across a dataframe.}
\usage{
count_missing(data)
}
\arguments{
\item{data}{A dataframe or tibble.}
}
\value{
A tibble.
}
\description{
A function to count missing values across a dataframe.
}
\examples{
guinea[[3]] |> count_missing()

}
