% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famd_groups.R
\name{famd_groups}
\alias{famd_groups}
\title{Form Groups Function}
\usage{
famd_groups(
  data,
  column = c("diam", "rim_angle", "simple_rim_type"),
  graph = FALSE
)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{column}{A character vector specifying the columns to be used for FAMD and clustering. Default columns are "diam", "rim_angle", and "simple_rim_type".}

\item{graph}{Logical. If TRUE, a cluster plot will be generated and printed. Default is FALSE.}
}
\value{
A list containing the following components:
\describe{
\item{complete_data}{A data frame with clustered data.}
\item{famd_input_data}{A data frame with input data used in FAMD.}
\item{results_cluster}{Results of HCPC clustering.}
\item{results_famd}{Results of FAMD analysis.}
}
}
\description{
This function performs various operations on the provided data, including FAMD (Factor Analysis of Mixed Data) and HCPC (Hierarchical Clustering on Principal Components) clustering. It also optionally generates and prints a cluster plot.
}
\examples{
\dontrun{
form_groups(data = my_data, column = c("diam", "rim_angle", "simple_rim_type"), graph = TRUE)
}

}
\seealso{
\code{\link{FactoMineR::FAMD}}
\code{\link{FactoMineR::HCPC}}
\code{\link{factoextra::fviz_cluster}}
}
