% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_letters.R
\name{get_first_letters}
\alias{get_first_letters}
\title{Get First Letters of Words in a String}
\usage{
get_first_letters(input_string)
}
\arguments{
\item{input_string}{The input string containing words.}
}
\value{
A character string with the first letters of words combined with "clust_" prefix.
}
\description{
This function takes an input string, splits it into words, and returns the first letters of each word combined as a single string with "clust_" prefix.
}
\examples{
get_first_letters("Hello World") # Returns "clust_HW"

}
