% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_summary}
\alias{get_summary}
\title{Get a summary list of all variables}
\usage{
get_summary(data)
}
\arguments{
\item{data}{A dataframe or tibble.}
}
\value{
A list
}
\description{
Get a summary list of all variables
}
\examples{
get_summary(guinea[[3]])
}
