% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guinea}
\alias{guinea}
\title{Pottery data from the Republic of Guinea}
\format{
\subsection{\code{guinea}}{

A list of three dataframes
\describe{
\item{rim}{rim sherd data}
\item{body}{body sherd data}
\item{all_sherds}{combined rim and body sherd data on shared fields}
...
}
}
}
\source{
excavations in Niandankoro, Republic of Guinea, 2023. Nikolas Gestrich
}
\usage{
guinea
}
\description{
Pottery data from the Republic of Guinea
}
\keyword{datasets}
