% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_numerics.R
\name{plot_numerics}
\alias{plot_numerics}
\title{Make a box and violin plot of all numeric columns in a dataset.}
\usage{
plot_numerics(data)
}
\arguments{
\item{data}{A dataframe or tibble.}
}
\value{
A box and violin plot
}
\description{
Make a box and violin plot of all numeric columns in a dataset.
}
\examples{
plot_numerics(guinea[[3]])

}
