% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_series.R
\name{plot_series}
\alias{plot_series}
\title{a wrapper function around tabula::plot_ford() in order to ease data
preparation}
\usage{
plot_series(data, context_var, attribute_var, context_order)
}
\arguments{
\item{data}{A dataframe or tibble.}

\item{context_var}{The variable that designates the context or assemblage}

\item{attribute_var}{The attribute that is the target for assessing seriation}

\item{context_order}{A vector giving the predefined order for the values in
context_var}
}
\value{
A ggplot object
}
\description{
a wrapper function around tabula::plot_ford() in order to ease data
preparation
}
\examples{
order <- as.character(c(1:26))
guinea[[1]] |> plot_series(context, rim_type, order)
}
