% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_summary}
\alias{plot_summary}
\title{get summary plots of all variables in a dataset}
\usage{
plot_summary(data)
}
\arguments{
\item{data}{A dataframe or tibble}
}
\value{
A list of plots
}
\description{
get summary plots of all variables in a dataset
}
\examples{
guinea[[3]] |> plot_summary()
}
