% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potteR-package.R
\docType{package}
\name{potteR-package}
\alias{potteR}
\alias{potteR-package}
\title{potteR: Analysis of Archaeological Pottery from Africa}
\description{
This package provides functions for the analysis of archaeological ceramics from Africa. It is intended to build up to a complete analytical workflow for artefacts recorded following the system proposed by Jesse & Nowotnick (2022).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nickgestrich/potteR}
  \item Report bugs at \url{https://github.com/nickgestrich/potteR/issues}
}

}
\author{
\strong{Maintainer}: Nikolas Gestrich \email{gestrich@uni-frankfurt.de} (\href{https://orcid.org/0000-0001-6376-1972}{ORCID})

}
\keyword{internal}
