% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary.R
\name{print_summary}
\alias{print_summary}
\title{Generate a pdf report on all the variables in a dataset}
\usage{
print_summary(data, file)
}
\arguments{
\item{data}{A dataframe or tibble.}

\item{file}{A filepath to a .pdf document}
}
\value{
A pdf document.
}
\description{
Generate a pdf report on all the variables in a dataset
}
\examples{
<<<<<<< HEAD
\dontrun{
guinea[[3]] |>  print_summary(file = here::here("summary.pdf"))
}
=======
guinea[[3]] |>  print_summary(file = here::here("summary.pdf"))

>>>>>>> bc46d271230e1bdfbe52534fdbc9bbc71dc98fc6
}
