% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_categoricals.R
\name{sum_categoricals}
\alias{sum_categoricals}
\title{Make a list of counts and percentages of all categorical vectors in a dataset.}
\usage{
sum_categoricals(data)
}
\arguments{
\item{data}{A dataframe or tibble.}
}
\value{
A list.
}
\description{
Make a list of counts and percentages of all categorical vectors in a dataset.
}
\examples{
guinea[[1]] |> sum_categoricals()
}
