% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_numerics.R
\name{sum_numerics}
\alias{sum_numerics}
\title{Get summary statistics for all numeric vectors in a dataset.}
\usage{
sum_numerics(data)
}
\arguments{
\item{data}{A dataframe or tibble}
}
\value{
A tibble.
}
\description{
Get summary statistics for all numeric vectors in a dataset.
}
\examples{
guinea[[1]] |> sum_numerics()
}
