% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_row_percent.R
\name{to_row_percent}
\alias{to_row_percent}
\title{Turn count data into percentages by row}
\usage{
to_row_percent(data)
}
\arguments{
\item{data}{a dataframe or tibble of count data}
}
\value{
a tibble of percentage data
}
\description{
Turn count data into percentages by row
}
\examples{
x <- tibble::tibble(x = c("a", "b", "c"), y = 4:6, z = 6:8)
to_row_percent(x)
x |> to_row_percent()
}
